# specs/005-blog-content-management/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Blog Content Management API Delta
  version: 1.0.0
  description: >-
    Private blog-module additions for feature 005. Existing public post and RSS
    contracts remain unchanged and are intentionally not expanded here.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /posts:
    get:
      summary: List owned posts with blog filters
      parameters:
        - { $ref: '#/components/parameters/SearchQuery' }
        - { $ref: '#/components/parameters/ContentClass' }
        - { $ref: '#/components/parameters/ContentTypeId' }
        - { $ref: '#/components/parameters/ContentStatus' }
        - { $ref: '#/components/parameters/AiDisplayStatus' }
        - { $ref: '#/components/parameters/SourceType' }
        - { $ref: '#/components/parameters/CategoryId' }
        - { $ref: '#/components/parameters/TagId' }
        - { $ref: '#/components/parameters/KeywordId' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: Cursor page of owned posts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
    post:
      summary: Create an empty owned draft and first revision
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostCreate' }
      responses:
        '201':
          description: Durably saved draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /posts/{post_id}/optimize:
    post:
      summary: Submit AI work bound to the current revision and fixed Skill version
      parameters: [{ $ref: '#/components/parameters/PostId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OptimizationRequest' }
      responses:
        '202':
          description: Durable job and run binding
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BlogJob' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /posts/{post_id}/candidates:
    get:
      summary: List AI candidates for an owned Post
      parameters:
        - { $ref: '#/components/parameters/PostId' }
        - in: query
          name: status
          schema: { type: string, enum: [pending, merge_required, applied, rejected, copied] }
      responses:
        '200':
          description: Candidate summaries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AiCandidate' }
  /posts/{post_id}/candidates/{candidate_id}:
    get:
      summary: Get three-way candidate comparison data
      parameters:
        - { $ref: '#/components/parameters/PostId' }
        - { $ref: '#/components/parameters/CandidateId' }
      responses:
        '200':
          description: Base, current and candidate snapshots with field diffs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CandidateComparison' }
        '404': { $ref: '#/components/responses/Problem' }
  /posts/{post_id}/candidates/{candidate_id}/decide:
    post:
      summary: Apply selected fields, reject, keep current or copy candidate
      parameters:
        - { $ref: '#/components/parameters/PostId' }
        - { $ref: '#/components/parameters/CandidateId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CandidateDecision' }
      responses:
        '200':
          description: Decision and optional new current revision
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CandidateDecisionResult' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /blog/triage:
    get:
      summary: List Posts derived into the triage work queue
      parameters:
        - { $ref: '#/components/parameters/SourceType' }
        - { $ref: '#/components/parameters/ContentClass' }
        - { $ref: '#/components/parameters/AiDisplayStatus' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: Triage Post page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
...
  /blog/content-types:
    get:
      summary: List owned enabled and historical content types
      responses:
        '200':
          description: Content types
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ContentType' }
    post:
      summary: Create a configurable content type
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentTypeWrite' }
      responses:
        '201':
          description: Created content type
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentType' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /blog/search:
    get:
      summary: Search owned blog title, body, code, source and structured fields
      parameters:
        - { $ref: '#/components/parameters/SearchQueryRequired' }
        - { $ref: '#/components/parameters/ContentClass' }
        - { $ref: '#/components/parameters/ContentTypeId' }
        - { $ref: '#/components/parameters/CategoryId' }
        - { $ref: '#/components/parameters/TagId' }
        - { $ref: '#/components/parameters/KeywordId' }
        - { $ref: '#/components/parameters/AiDisplayStatus' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: Search page with match fields and safe highlights
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchPage' }
  /blog/timeline:
    get:
      summary: Return owned articles grouped by occurrence or creation time
      parameters:
        - in: query
          name: time_basis
          schema: { type: string, enum: [occurred_at, created_at], default: occurred_at }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - { $ref: '#/components/parameters/ContentClass' }
        - { $ref: '#/components/parameters/ContentTypeId' }
        - { $ref: '#/components/parameters/CategoryId' }
        - { $ref: '#/components/parameters/TagId' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: Timeline page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimelinePage' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
  parameters:
    PostId: { in: path, name: post_id, required: true, schema: { type: string, format: uuid } }
...
    CandidateId: { in: path, name: candidate_id, required: true, schema: { type: string, format: uuid } }
...
    SearchQuery: { in: query, name: q, schema: { type: string, maxLength: 500 } }
    SearchQueryRequired: { in: query, name: q, required: true, schema: { type: string, minLength: 1, maxLength: 500 } }
    ContentClass: { in: query, name: content_class, schema: { type: string, maxLength: 32 } }
    ContentTypeId: { in: query, name: content_type_id, schema: { type: string, format: uuid } }
    ContentStatus: { in: query, name: content_status, schema: { type: string, maxLength: 32 } }
    AiDisplayStatus: { in: query, name: ai_status, schema: { type: string, maxLength: 32 } }
    SourceType: { in: query, name: source_type, schema: { type: string, maxLength: 24 } }
    CategoryId: { in: query, name: category_id, schema: { type: string, format: uuid } }
    TagId: { in: query, name: tag_id, schema: { type: string, format: uuid } }
    KeywordId: { in: query, name: keyword_id, schema: { type: string, format: uuid } }
    Cursor: { in: query, name: cursor, schema: { type: string, maxLength: 500 } }
  responses:
    Problem:
      description: RFC 9457 problem details; never contains private raw content
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Post:
      type: object
      additionalProperties: false
      required: [id, title, markdown, status, content_status, content_class, language, structured_data, version, current_revision_id, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string, maxLength: 240 }
        subtitle: { type: [string, 'null'], maxLength: 240 }
        summary: { type: [string, 'null'], maxLength: 2000 }
        markdown: { type: string, maxLength: 200000 }
        status: { type: string, enum: [draft, private, published] }
        content_status: { $ref: '#/components/schemas/ContentStatus' }
        content_class: { type: string, maxLength: 32 }
        content_type_id: { type: [string, 'null'], format: uuid }
        category_id: { type: [string, 'null'], format: uuid }
        tag_ids: { type: array, items: { type: string, format: uuid }, maxItems: 50 }
        keyword_ids: { type: array, items: { type: string, format: uuid }, maxItems: 100 }
        language: { type: string, maxLength: 16 }
        editor_mode: { type: string, enum: [markdown, rich, split] }
        occurred_at: { type: [string, 'null'], format: date-time }
        location: { type: [string, 'null'], maxLength: 240 }
        project: { type: [string, 'null'], maxLength: 240 }
        structured_data: { type: object, maxProperties: 100 }
        source_summary: { type: array, items: { $ref: '#/components/schemas/PostSourceSummary' } }
        ai_summary: { $ref: '#/components/schemas/PostAiSummary' }
        version: { type: integer, minimum: 1 }
        current_revision_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PostCreate:
      type: object
      additionalProperties: false
      properties:
        title: { type: string, default: 未命名文章, maxLength: 240 }
        markdown: { type: string, default: '', maxLength: 200000 }
        content_class: { type: string, default: essay }
        content_type_id: { type: [string, 'null'], format: uuid }
        ai_enabled: { type: boolean, default: false }
        skill_id: { type: [string, 'null'], format: uuid }
        save_as_defaults: { type: boolean, default: false }
...
    ContentStatus:
      type: string
      enum: [pending_capture, pending_parse, triage, draft, ai_queued, ai_processing, ai_review, merge_required, completed, archived, discarded]
...
    PostSourceSummary:
      type: object
      required: [id, source_type, status, captured_at]
      properties:
        id: { type: string, format: uuid }
        source_type: { type: string }
        status: { type: string }
        original_url: { type: [string, 'null'], format: uri }
        original_title: { type: [string, 'null'] }
        captured_at: { type: string, format: date-time }
...
    OptimizationRequest:
      type: object
      additionalProperties: false
      required: [post_version, optimization_type]
      properties:
        post_version: { type: integer, minimum: 1 }
        optimization_type: { type: string, enum: [full, language, structure, metadata, check, reoptimize] }
        scope: { type: string, enum: [all, body, metadata, selected_fields], default: all }
        selected_fields: { type: array, maxItems: 200, uniqueItems: true, items: { type: string, maxLength: 128 } }
        skill_id: { type: [string, 'null'], format: uuid }
        provider_key: { type: [string, 'null'], enum: [radio, aiassist, null] }
        model_key: { type: [string, 'null'], maxLength: 120 }
        instruction: { type: [string, 'null'], maxLength: 2000 }
        request_nonce: { type: [string, 'null'], maxLength: 64 }
    AiCandidate:
      type: object
      required: [id, post_id, run_id, base_revision_id, candidate_revision_id, status, validation, created_at]
      properties:
        id: { type: string, format: uuid }
        post_id: { type: string, format: uuid }
        run_id: { type: string, format: uuid }
        base_revision_id: { type: string, format: uuid }
        candidate_revision_id: { type: string, format: uuid }
        status: { type: string, enum: [pending, merge_required, applied, rejected, copied] }
        field_diff: { type: object }
        validation: { type: object }
        created_at: { type: string, format: date-time }
        reviewed_at: { type: [string, 'null'], format: date-time }
    CandidateComparison:
      type: object
      required: [candidate, base, current, proposed, body_diff, field_diffs]
      properties:
        candidate: { $ref: '#/components/schemas/AiCandidate' }
        base: { $ref: '#/components/schemas/PostRevision' }
        current: { $ref: '#/components/schemas/PostRevision' }
        proposed: { $ref: '#/components/schemas/PostRevision' }
        body_diff: { type: string }
        field_diffs: { type: array, items: { $ref: '#/components/schemas/FieldDiff' } }
    CandidateDecision:
      type: object
      additionalProperties: false
      required: [action, current_post_version]
      properties:
        action: { type: string, enum: [apply_all, apply_body, apply_metadata, apply_fields, keep_current, reject, copy] }
        current_post_version: { type: integer, minimum: 1 }
        selected_fields: { type: array, maxItems: 200, uniqueItems: true, items: { type: string, maxLength: 128 } }
        rejected_fields: { type: array, maxItems: 200, uniqueItems: true, items: { type: string, maxLength: 128 } }
    CandidateDecisionResult:
      type: object
      required: [candidate, post, result_revision_id]
      properties:
        candidate: { $ref: '#/components/schemas/AiCandidate' }
        post: { oneOf: [{ $ref: '#/components/schemas/Post' }, { type: 'null' }] }
        result_revision_id: { type: [string, 'null'], format: uuid }
    PostRevision:
      type: object
      required: [id, post_id, revision_number, source, markdown, snapshot, schema_version, created_at]
      properties:
        id: { type: string, format: uuid }
        post_id: { type: string, format: uuid }
        revision_number: { type: integer, minimum: 1 }
        parent_revision_id: { type: [string, 'null'], format: uuid }
        base_revision_id: { type: [string, 'null'], format: uuid }
        source: { type: string, enum: [capture, user_edit, ai_candidate, ai_applied, restore, import, merge] }
        markdown: { type: string, maxLength: 200000 }
        snapshot: { type: object }
        change_summary: { type: [string, 'null'], maxLength: 500 }
        async_job_id: { type: [string, 'null'], format: uuid }
        skill_version_id: { type: [string, 'null'], format: uuid }
        schema_version: { type: string }
        applied_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
...
    FieldDiff:
      type: object
      required: [path, before, after, policy, validation]
      properties:
        path: { type: string, maxLength: 128 }
        before: {}
        after: {}
        policy: { type: string }
        validation: { type: object }
    PostAiSummary:
      type: object
      required: [display_status, optimization_count]
      properties:
        display_status: { type: [string, 'null'] }
        optimization_count: { type: integer, minimum: 0 }
        first_optimized_at: { type: [string, 'null'], format: date-time }
        last_optimized_at: { type: [string, 'null'], format: date-time }
        latest_job_id: { type: [string, 'null'], format: uuid }
        pending_candidate_id: { type: [string, 'null'], format: uuid }
    BlogJob:
      type: object
      required: [id, job_type, status, display_status, progress, retry_count, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        job_type: { type: string }
        entity_type: { type: [string, 'null'] }
        entity_id: { type: [string, 'null'], format: uuid }
        status: { type: string, enum: [pending, queued, processing, waiting_user, completed, failed, cancelled] }
        display_status: { type: string }
        current_stage: { type: [string, 'null'] }
        progress: { type: integer, minimum: 0, maximum: 100 }
        retry_count: { type: integer, minimum: 0 }
        result: { type: [object, 'null'] }
        error: { oneOf: [{ $ref: '#/components/schemas/JobError' }, { type: 'null' }] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
    JobError:
      type: object
      required: [code, message, retryable]
      properties:
        code: { type: string, maxLength: 64 }
        message: { type: string, maxLength: 500 }
        retryable: { type: boolean }
...
    ContentType:
      type: object
      additionalProperties: false
      required: [id, content_class, key, name, field_schema, enabled, schema_version, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        content_class: { type: string, maxLength: 32 }
        key: { type: string, pattern: '^[a-z][a-z0-9_-]{0,63}$' }
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: [string, 'null'], maxLength: 500 }
        field_schema: { type: object, maxProperties: 100 }
        sort_order: { type: integer, default: 0 }
        enabled: { type: boolean }
        schema_version: { type: integer, minimum: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ContentTypeWrite:
      type: object
      additionalProperties: false
      required: [content_class, key, name, field_schema, enabled]
      properties:
        content_class: { type: string, maxLength: 32 }
        key: { type: string, pattern: '^[a-z][a-z0-9_-]{0,63}$' }
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: [string, 'null'], maxLength: 500 }
        field_schema: { type: object, maxProperties: 100 }
        sort_order: { type: integer, default: 0 }
        enabled: { type: boolean }
...
    PostPage:
      type: object
      required: [items, next_cursor, applied_filters]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Post' } }
        next_cursor: { type: [string, 'null'] }
        applied_filters: { type: object }
    SearchPage:
      type: object
      required: [query, items, next_cursor, applied_filters, index_pending_count]
      properties:
        query: { type: string }
        items:
          type: array
          items:
            type: object
            required: [post, matched_fields, highlights]
            properties:
              post: { $ref: '#/components/schemas/Post' }
              matched_fields: { type: array, items: { type: string } }
              highlights: { type: array, items: { type: string } }
        next_cursor: { type: [string, 'null'] }
        applied_filters: { type: object }
        index_pending_count: { type: integer, minimum: 0 }
    TimelinePage:
      type: object
      required: [items, next_cursor, time_basis]
      properties:
        items:
          type: array
          items:
            type: object
            required: [post, display_time, time_basis]
            properties:
              post: { $ref: '#/components/schemas/Post' }
              display_time: { type: string, format: date-time }
              time_basis: { type: string, enum: [occurred_at, created_at] }
        next_cursor: { type: [string, 'null'] }
        time_basis: { type: string, enum: [occurred_at, created_at] }
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        trace_id: { type: string }